Remove old DetailList
- use the new variant for translations

RISDEV-12108
